# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "astroimage",
    "version": "0.1.0"
  },
  "paths": {
...
    "/fits/metadata": {
      "post": {
        "tags": [
          "fits"
        ],
        "summary": "Extract Fits Metadata",
        "operationId": "extractFitsMetadata",
        "parameters": [
          {
            "name": "hdu",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional image HDU index; defaults to the primary 2D image HDU",
              "title": "Hdu"
            },
            "description": "Optional image HDU index; defaults to the primary 2D image HDU"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extractFitsMetadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitsMetadataSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/image": {
      "get": {
        "tags": [
          "hub"
        ],
        "summary": "List Astro Images",
        "operationId": "listHubbleImages",
        "parameters": [
          {
            "name": "cuerpo_celeste",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cuerpo Celeste"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRecordsResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/image/search": {
      "get": {
        "tags": [
          "hub"
        ],
        "summary": "Search Astro Image",
        "operationId": "fetchHubbleImage",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Celestial body to fetch from Hubble",
              "title": "Query"
            },
            "description": "Celestial body to fetch from Hubble"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchImageResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/image/{record_id}/info": {
      "get": {
        "tags": [
          "hub"
        ],
        "summary": "Get Image Info",
        "operationId": "getImageInfo",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Stored FITS record id",
              "title": "Record Id"
            },
            "description": "Stored FITS record id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitsMetadataSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/image/{record_id}/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Detect Sources",
        "operationId": "detectSources",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Stored FITS record id",
              "title": "Record Id"
            },
            "description": "Stored FITS record id"
          },
          {
            "name": "hdu",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional image HDU index; defaults to the first 2D image HDU",
              "title": "Hdu"
            },
            "description": "Optional image HDU index; defaults to the first 2D image HDU"
          },
          {
            "name": "fwhm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0.5,
              "default": 5.5,
              "title": "Fwhm"
            }
          },
          {
            "name": "sigma",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1.0,
              "default": 9.0,
              "title": "Sigma"
            }
          },
          {
            "name": "min_snr",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0.0,
              "default": 6.0,
              "title": "Min Snr"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 0.18,
              "title": "Min Score"
            }
          },
          {
            "name": "min_distance",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0.0,
              "default": 4.0,
              "title": "Min Distance"
            }
          },
          {
            "name": "visual_weight",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 0.8,
              "title": "Visual Weight"
            }
          },
          {
            "name": "visual_area_radius",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1.0,
              "default": 7.0,
              "title": "Visual Area Radius"
            }
          },
          {
            "name": "visual_area_sigma",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0.0,
              "default": 2.0,
              "title": "Visual Area Sigma"
            }
          },
          {
            "name": "max_sources",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 50,
              "title": "Max Sources"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDetectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_extractFitsMetadata": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "FITS file to inspect"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_extractFitsMetadata"
      },
      "DetectionSummarySchema": {
        "properties": {
          "point_count": {
            "type": "integer",
            "title": "Point Count",
            "default": 0
          },
          "extended_count": {
            "type": "integer",
            "title": "Extended Count",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DetectionSummarySchema"
      },
      "ExtendedSourceSchema": {
        "properties": {
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "object_type": {
            "type": "string",
            "const": "extended",
            "title": "Object Type",
            "default": "extended"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_id",
          "rank"
        ],
        "title": "ExtendedSourceSchema"
      },
      "FetchImageResponseSchema": {
        "properties": {
          "record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Record Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record_id"
        ],
        "title": "FetchImageResponseSchema"
      },
...
      "FitsHduInfoSchema": {
        "properties": {
          "selected": {
            "type": "integer",
            "title": "Selected"
          },
          "image_indices": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Image Indices"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/FitsHduDetailSchema"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "selected"
        ],
        "title": "FitsHduInfoSchema"
      },
      "FitsImageInfoSchema": {
        "properties": {
          "shape": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shape"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "datamin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datamin"
          },
          "datamax": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datamax"
          },
          "datamean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datamean"
          },
          "median": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median"
          },
          "background": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FitsImageInfoSchema"
      },
      "FitsInstrumentInfoSchema": {
        "properties": {
          "telescope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telescope"
          },
          "instrument": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument"
          },
          "detector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detector"
          },
          "filter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filter Name"
          },
          "exptime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exptime"
          },
          "date_obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Obs"
          },
          "time_obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Obs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FitsInstrumentInfoSchema"
      },
      "FitsMetadataSchema": {
        "properties": {
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "image": {
            "$ref": "#/components/schemas/FitsImageInfoSchema"
          },
          "instrument": {
            "$ref": "#/components/schemas/FitsInstrumentInfoSchema"
          },
          "photometry": {
            "$ref": "#/components/schemas/FitsPhotometryInfoSchema"
          },
          "wcs": {
            "$ref": "#/components/schemas/FitsWcsInfoSchema"
          },
          "hdus": {
            "$ref": "#/components/schemas/FitsHduInfoSchema"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/FitsTableInfoSchema"
            },
            "type": "array",
            "title": "Tables"
          },
          "header": {
            "additionalProperties": true,
            "type": "object",
            "title": "Header"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hdus"
        ],
        "title": "FitsMetadataSchema"
      },
      "FitsPhotometryInfoSchema": {
        "properties": {
          "photflam": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photflam"
          },
          "photplam": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photplam"
          },
          "photbw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photbw"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FitsPhotometryInfoSchema"
      },
      "FitsRecordSummarySchema": {
        "properties": {
          "record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Record Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record_id",
          "name"
        ],
        "title": "FitsRecordSummarySchema"
      },
      "FitsTableInfoSchema": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "index",
          "name",
          "rows",
          "columns"
        ],
        "title": "FitsTableInfoSchema"
      },
      "FitsWcsInfoSchema": {
        "properties": {
          "present": {
            "type": "boolean",
            "title": "Present",
            "default": false
          },
          "naxis": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Naxis"
          },
          "crval": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crval"
          },
          "crpix": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crpix"
          },
          "ctype": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctype"
          },
          "cunit": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cunit"
          },
          "cd": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cd"
          },
          "cdelt": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cdelt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FitsWcsInfoSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListRecordsResponseSchema": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/FitsRecordSummarySchema"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "records"
        ],
        "title": "ListRecordsResponseSchema"
      },
      "PointSourceSchema": {
        "properties": {
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "xcentroid": {
            "type": "number",
            "title": "Xcentroid"
          },
          "ycentroid": {
            "type": "number",
            "title": "Ycentroid"
          },
          "snr": {
            "type": "number",
            "title": "Snr"
          },
          "relevance_score": {
            "type": "number",
            "title": "Relevance Score"
          },
          "peak": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak"
          },
          "flux": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flux"
          },
          "object_type": {
            "type": "string",
            "const": "point",
            "title": "Object Type",
            "default": "point"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_id",
          "rank",
          "xcentroid",
          "ycentroid",
          "snr",
          "relevance_score"
        ],
        "title": "PointSourceSchema"
      },
      "SourceDetectionResponse": {
        "properties": {
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "summary": {
            "$ref": "#/components/schemas/DetectionSummarySchema"
          },
          "point_sources": {
            "items": {
              "$ref": "#/components/schemas/PointSourceSchema"
            },
            "type": "array",
            "title": "Point Sources"
          },
          "extended_sources": {
            "items": {
              "$ref": "#/components/schemas/ExtendedSourceSchema"
            },
            "type": "array",
            "title": "Extended Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "SourceDetectionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}